requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
plotly
cohere
python-dotenv